Extract shared haversine utility from MapView and lookup
Consolidate identical toRadians() and haversineMiles() definitions from
components/MapView.tsx and lib/cpt/lookup.ts into a shared lib/geo.ts.
Eliminates 45 lines of duplicated pure math with a single 13-line module.

Closes #67
