+++ components/MapView.tsx
@@ -4,6 +4,7 @@ import { useEffect, useRef, useState } from "react";
 import { setOptions, importLibrary } from "@googlemaps/js-api-loader";
 import type { ChargeResult } from "@/types";
 import { getDisplayPrice, formatDisplayPrice } from "@/lib/format";
+import { haversineMiles } from "@/lib/geo";
 
 interface MapViewProps {
   results: ChargeResult[];
@@ -78,30 +79,6 @@ const FIT_MARKER_LIMIT = 24;
 const FIT_RADIUS_MILES = 30;
 const FIT_MIN_ZOOM = 10;
 const FIT_MAX_ZOOM = 13;
-
-function toRadians(degrees: number): number {
-  return degrees * (Math.PI / 180);
-}
-
-function haversineMiles(
-  originLat: number,
-  originLng: number,
-  targetLat: number,
-  targetLng: number
-): number {
-  const earthRadiusMiles = 3958.8;
-  const latDelta = toRadians(targetLat - originLat);
-  const lngDelta = toRadians(targetLng - originLng);
-
-  const a =
-    Math.sin(latDelta / 2) ** 2 +
-    Math.cos(toRadians(originLat)) *
-      Math.cos(toRadians(targetLat)) *
-      Math.sin(lngDelta / 2) ** 2;
-  const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-
-  return earthRadiusMiles * c;
-}
 
 export function MapView({
   results,
