+++ lib/cpt/lookup.ts
@@ -1,5 +1,6 @@
 import { createClient } from "@/lib/supabase/server";
 import { milesToKm, kmToMiles } from "@/lib/units";
+import { haversineMiles } from "@/lib/geo";
 // @ts-expect-error zipcodes has no bundled TypeScript declarations.
 import zipcodes from "zipcodes";
 import type {
@@ -138,30 +139,6 @@ export function createLookupDiagnostics(): LookupDiagnostics {
     fallbackExhausted: false,
     stageSummaries: [],
   };
-}
-
-function toRadians(degrees: number): number {
-  return degrees * (Math.PI / 180);
-}
-
-function haversineMiles(
-  originLat: number,
-  originLng: number,
-  targetLat: number,
-  targetLng: number
-): number {
-  const earthRadiusMiles = 3958.8;
-  const latDelta = toRadians(targetLat - originLat);
-  const lngDelta = toRadians(targetLng - originLng);
-
-  const a =
-    Math.sin(latDelta / 2) ** 2 +
-    Math.cos(toRadians(originLat)) *
-      Math.cos(toRadians(targetLat)) *
-      Math.sin(lngDelta / 2) ** 2;
-
-  const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-  return earthRadiusMiles * c;
 }
 
 function escapeRegex(value: string): string {
